#59 Refactor exception-handler -> commons-spring-boot-starter
