+++ services/commons-spring-boot-starter/settings.gradle
@@ -1,0 +1,1 @@
+rootProject.name = 'commons-spring-boot-starter'
