+++ services/commons-spring-boot-starter/src/main/java/com/senla/worklog/reminder/exception/UnexpectedApplicationException.java
@@ -1,0 +1,18 @@
+package com.senla.worklog.reminder.exception;
+
+
+import com.senla.worklog.reminder.exception.wrapper.ExceptionWrapper;
+
+/**
+ * This class represents an exception that is thrown when exception in
+ * application service was not wrapped by specific {@link ExceptionWrapper} into corresponding {@link ApplicationException}
+ */
+public class UnexpectedApplicationException extends ApplicationException {
+    public UnexpectedApplicationException(Throwable cause) {
+        super("Unexpected application error", cause);
+    }
+
+    public UnexpectedApplicationException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
