+++ services/commons-spring-boot-starter/src/main/java/com/senla/worklog/reminder/exception/wrapper/ExceptionWrapper.java
@@ -1,0 +1,24 @@
+package com.senla.worklog.reminder.exception.wrapper;
+
+import com.senla.worklog.reminder.exception.ApplicationException;
+
+/**
+ * An interface for classes that wrap exceptions in application-specific exceptions.
+ */
+public interface ExceptionWrapper {
+
+    /**
+     * Wraps the given exception in a specific {@link ApplicationException}.
+     *
+     * @param e the exception to wrap
+     * @return the wrapped exception
+     */
+    ApplicationException wrapInApplicationException(Exception e);
+
+    /**
+     * Gets the class of the exception type that this wrapper handles.
+     *
+     * @return the class of the handled exception type
+     */
+    Class<? extends Exception> getExceptionType();
+}
