+++ services/commons-spring-boot-starter/src/main/resources/application.yaml
