+++ services/employee/build.gradle
@@ -22,7 +22,7 @@ dependencies {
 	implementation 'org.liquibase:liquibase-core'
 	implementation 'org.mapstruct:mapstruct:1.5.3.Final'
 	implementation 'org.apache.httpcomponents:httpclient'
-	implementation 'com.senla.common:exception-handler:2.7.8'
+	implementation 'com.senla.worklog.reminder:commons-spring-boot-starter:2.7.8'
 	compileOnly 'org.projectlombok:lombok'
 	runtimeOnly 'org.postgresql:postgresql'
 	annotationProcessor 'org.projectlombok:lombok'
