+++ services/employee/src/test/java/com/senla/worklog/reminder/employee/adapter/in/rest/exception/handler/RestAdapterExceptionHandlerTest.java
@@ -1,18 +1,16 @@
 package com.senla.worklog.reminder.employee.adapter.in.rest.exception.handler;
 
-import com.senla.common.exception.UnexpectedApplicationException;
-import com.senla.common.exception.handler.model.AttributeApiSubError;
 import com.senla.worklog.reminder.employee.application.exception.ResourceNotFoundException;
 import com.senla.worklog.reminder.employee.application.exception.UniqueConstraintViolationException;
 import com.senla.worklog.reminder.employee.domain.exception.DuplicateAttributeException;
 import com.senla.worklog.reminder.employee.domain.exception.EmployeeNotFoundException;
+import com.senla.worklog.reminder.exception.handler.model.AttributeApiSubError;
 import org.junit.jupiter.api.Test;
 
 import static org.hibernate.validator.internal.util.Contracts.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.springframework.http.HttpStatus.BAD_REQUEST;
-import static org.springframework.http.HttpStatus.INTERNAL_SERVER_ERROR;
 import static org.springframework.http.HttpStatus.NOT_FOUND;
 
 class RestAdapterExceptionHandlerTest {
@@ -64,17 +62,5 @@ class RestAdapterExceptionHandlerTest {
         var subError = response.getBody().getErrors().get(0);
         assertEquals("username", ((AttributeApiSubError) subError).getAttributeName());
         assertEquals("john", ((AttributeApiSubError) subError).getAttributeValue());
-    }
-
-    @Test
-    void handleUnexpectedApplicationException_shouldReturn500_whenInputIsUnexpectedApplicationException() {
-        var cause = new CloneNotSupportedException();
-        var exception = new UnexpectedApplicationException(cause);
-
-        var responseEntity = exceptionHandler.handleUnexpectedApplicationException(exception);
-
-        assertEquals(responseEntity.getStatusCode(), INTERNAL_SERVER_ERROR);
-        assertNotNull(responseEntity.getBody());
-        assertEquals(responseEntity.getBody().getStatus(), 500);
     }
 }
