--- services/exception-handler/build.gradle
@@ -1,52 +1,0 @@
-import org.springframework.boot.gradle.plugin.SpringBootPlugin
-
-plugins {
-    id 'java'
-    id 'maven-publish'
-    id 'org.springframework.boot' version '2.7.8' apply false
-    id 'io.spring.dependency-management' version '1.0.15.RELEASE'
-}
-
-group = 'com.senla.common'
-version = '2.7.8'
-sourceCompatibility = '11'
-
-repositories {
-    mavenCentral()
-}
-
-publishing {
-    publications {
-        mavenJava(MavenPublication) {
-            from components.java
-
-            dependencies {
-                implementation 'org.springframework.boot:spring-boot-starter-web:2.7.8'
-                implementation 'org.springframework.boot:spring-boot-starter-aop:2.7.8'
-                implementation 'org.springframework.boot:spring-boot-starter-validation:2.7.8'
-            }
-        }
-    }
-    repositories {
-        mavenLocal()
-    }
-}
-
-dependencyManagement {
-    imports {
-        mavenBom SpringBootPlugin.BOM_COORDINATES
-    }
-}
-
-dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-    implementation 'org.springframework.boot:spring-boot-starter-aop'
-    implementation 'org.springframework.boot:spring-boot-starter-validation'
-    compileOnly 'org.projectlombok:lombok'
-    annotationProcessor 'org.projectlombok:lombok'
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
-}
-
-tasks.named('test') {
-    useJUnitPlatform()
-}
