--- services/exception-handler/src/main/java/com/senla/common/exception/DomainException.java
@@ -1,15 +1,0 @@
-package com.senla.common.exception;
-
-/**
- * This class is an abstract class that represents an exception that occurs within the domain layer of the application
- */
-public abstract class DomainException extends RuntimeException {
-
-    public DomainException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public DomainException(String message) {
-        super(message);
-    }
-}
