--- services/exception-handler/src/main/java/com/senla/common/exception/UnexpectedApplicationException.java
@@ -1,18 +1,0 @@
-package com.senla.common.exception;
-
-
-import com.senla.common.exception.wrapper.ExceptionWrapper;
-
-/**
- * This class represents an exception that is thrown when exception in
- * application service was not wrapped by specific {@link ExceptionWrapper} into corresponding {@link ApplicationException}
- */
-public class UnexpectedApplicationException extends ApplicationException {
-    public UnexpectedApplicationException(Throwable cause) {
-        super("Unexpected application error", cause);
-    }
-
-    public UnexpectedApplicationException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
