--- services/exception-handler/src/main/java/com/senla/common/exception/wrapper/ExceptionWrapper.java
@@ -1,24 +1,0 @@
-package com.senla.common.exception.wrapper;
-
-import com.senla.common.exception.ApplicationException;
-
-/**
- * An interface for classes that wrap exceptions in application-specific exceptions.
- */
-public interface ExceptionWrapper {
-
-    /**
-     * Wraps the given exception in a specific {@link ApplicationException}.
-     *
-     * @param e the exception to wrap
-     * @return the wrapped exception
-     */
-    ApplicationException wrapInApplicationException(Exception e);
-
-    /**
-     * Gets the class of the exception type that this wrapper handles.
-     *
-     * @return the class of the handled exception type
-     */
-    Class<? extends Exception> getExceptionType();
-}
