--- services/exception-handler/src/main/java/com/senla/common/exception/wrapper/ExceptionWrapperAutoConfiguration.java
@@ -1,21 +1,0 @@
-package com.senla.common.exception.wrapper;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.EnableAspectJAutoProxy;
-
-import java.util.List;
-
-@Configuration
-@EnableAspectJAutoProxy
-public class ExceptionWrapperAutoConfiguration {
-    @Bean
-    public ExceptionWrapperRegistry exceptionWrapperRegistry(List<ExceptionWrapper> exceptionWrappers) {
-        return new ExceptionWrapperRegistry(exceptionWrappers);
-    }
-
-    @Bean
-    public ExceptionWrapperAspect exceptionWrapperAspect(ExceptionWrapperRegistry exceptionWrapperRegistry) {
-        return new ExceptionWrapperAspect(exceptionWrapperRegistry);
-    }
-}
