--- services/exception-handler/src/main/resources/META-INF/spring.factories
@@ -1,3 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-  com.senla.common.exception.handler.CommonExceptionHandlerAdvice,\
-  com.senla.common.exception.wrapper.ExceptionWrapperAutoConfiguration
